+++ database/migrations/2026_04_09_033922_remove_endereco_id_from_pessoa_table.php
@@ -11,10 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('pessoa', function (Blueprint $table) {
-            $table->dropForeign('pessoas_endereco_id_foreign');
-            $table->dropColumn('endereco_id');
-        });
+        if (config('database.default') !== 'sqlite') {
+            Schema::table('pessoa', function (Blueprint $table) {
+                $table->dropForeign('pessoas_endereco_id_foreign');
+                $table->dropColumn('endereco_id');
+            });
+        }
     }
 
     /**
